learn

+++ Scripts/Handles.sh
@@ -83,7 +83,10 @@ if [ -f "$FW_FILE" ]; then
 fi
 
 #替换成最新版的golang
-cd $REPO_PATCH
-rm -rf feeds/packages/lang/golang
-git clone https://github.com/sbwml/packages_lang_golang -b 23.x feeds/packages/lang/golang
-cd $PKG_PATCH && echo "golang update done!"
+if [[ $WRT_REPO != *"lede"* ]]; then
+	cd $REPO_PATCH
+	rm -rf feeds/packages/lang/golang
+	git clone https://github.com/sbwml/packages_lang_golang -b 23.x feeds/packages/lang/golang
+	cd $PKG_PATCH && echo "golang update done!"
+fi
+
